+++ services/batch/mgmt/2021-06-01/batch/CHANGELOG.md
@@ -1,2 +1,33 @@
-# Change History
+# Unreleased
 
+## Breaking Changes
+
+### Removed Funcs
+
+1. *AccountCreateFuture.UnmarshalJSON([]byte) error
+1. AccountClient.Create(context.Context, string, string, AccountCreateParameters) (AccountCreateFuture, error)
+1. AccountClient.CreatePreparer(context.Context, string, string, AccountCreateParameters) (*http.Request, error)
+1. AccountClient.CreateResponder(*http.Response) (Account, error)
+1. AccountClient.CreateSender(*http.Request) (AccountCreateFuture, error)
+
+### Struct Changes
+
+#### Removed Structs
+
+1. AccountCreateFuture
+
+## Additive Changes
+
+### New Funcs
+
+1. *AccountCreateABCFuture.UnmarshalJSON([]byte) error
+1. AccountClient.CreateABC(context.Context, string, string, AccountCreateParameters) (AccountCreateABCFuture, error)
+1. AccountClient.CreateABCPreparer(context.Context, string, string, AccountCreateParameters) (*http.Request, error)
+1. AccountClient.CreateABCResponder(*http.Response) (Account, error)
+1. AccountClient.CreateABCSender(*http.Request) (AccountCreateABCFuture, error)
+
+### Struct Changes
+
+#### New Structs
+
+1. AccountCreateABCFuture